requirements.txt
google-generativeai
kaggle
pandas
termcolor
python-dotenv